add notifications sign to mango create account process (#254)
* ledger setup + lib bump

* fix

* remove ledger checkbox

* fix

* update layout/copy

* fix typo

---------
