+++ components/modals/MangoAccountsListModal.tsx
@@ -102,7 +102,7 @@ const MangoAccountsListModal = ({
   return (
     <Modal isOpen={isOpen} onClose={onClose}>
       <div className="inline-block w-full transform overflow-x-hidden">
-        <div className="flex min-h-[324px] flex-col justify-between">
+        <div className="flex min-h-[364px] flex-col justify-between">
           <div>
             <h2 className="text-center">{t('accounts')}</h2>
             {loading ? (
