+++ components/modals/UserSetupModal.tsx
@@ -22,7 +22,7 @@ import {
   useMemo,
   useState,
 } from 'react'
-import { notify } from 'utils/notifications'
+import { createSolanaMessage, notify } from 'utils/notifications'
 import ActionTokenList from '../account/ActionTokenList'
 import ButtonGroup from '../forms/ButtonGroup'
 import Input from '../forms/Input'
@@ -31,7 +31,6 @@ import Label from '../forms/Label'
 // import EditNftProfilePic from '../profile/EditNftProfilePic'
 // import EditProfileForm from '../profile/EditProfileForm'
 import Button, { LinkButton } from '../shared/Button'
-import InlineNotification from '../shared/InlineNotification'
 import Loading from '../shared/Loading'
 import MaxAmountButton from '../shared/MaxAmountButton'
 import SolBalanceWarnings from '../shared/SolBalanceWarnings'
@@ -46,6 +45,8 @@ import useLocalStorageState from 'hooks/useLocalStorageState'
 import { ACCEPT_TERMS_KEY, MAX_ACCOUNTS } from 'utils/constants'
 import { ACCOUNT_ACTIONS_NUMBER_FORMAT_CLASSES } from '@components/BorrowForm'
 import { WalletReadyState } from '@solana/wallet-adapter-base'
+import Switch from '@components/forms/Switch'
+import NotificationCookieStore from '@store/notificationCookieStore'
 
 const UserSetupModal = ({
   isOpen,
@@ -65,6 +66,7 @@ const UserSetupModal = ({
   const [depositAmount, setDepositAmount] = useState('')
   const [submitDeposit, setSubmitDeposit] = useState(false)
   const [sizePercentage, setSizePercentage] = useState('')
+  const [signToNotifications, setSignToNotifications] = useState(true)
   // const [showEditProfilePic, setShowEditProfilePic] = useState(false)
   const { maxSolDeposit } = useSolBalance()
   const banks = useBanksWithBalances('walletBalance')
@@ -72,6 +74,9 @@ const UserSetupModal = ({
   const [walletsToDisplay, setWalletstoDisplay] = useState<'default' | 'all'>(
     'default',
   )
+  //used to sign txes
+  const walletContext = useWallet()
+  const setCookie = NotificationCookieStore((s) => s.setCookie)
 
   const walletsDisplayed = useMemo(() => {
     const firstFive = wallets.slice(0, 5)
@@ -114,6 +119,9 @@ const UserSetupModal = ({
       )
       actions.fetchMangoAccounts(publicKey)
       if (tx) {
+        if (signToNotifications) {
+          createSolanaMessage(walletContext, setCookie)
+        }
         actions.fetchWalletTokens(publicKey) // need to update sol balance after account rent
         setShowSetupStep(3)
         notify({
@@ -357,9 +365,7 @@ const UserSetupModal = ({
                   <h2 className="mb-4 font-display text-3xl tracking-normal md:text-5xl lg:text-6xl">
                     {t('onboarding:create-account')}
                   </h2>
-                  <p className="text-base">
-                    {t('onboarding:create-account-desc')}
-                  </p>
+                  <p className="text-base">{t('insufficient-sol')}</p>
                 </div>
                 <div className="mb-4">
                   <Label text={t('account-name')} optional />
@@ -376,11 +382,18 @@ const UserSetupModal = ({
                   />
                 </div>
                 <SolBalanceWarnings className="mt-4" />
-                <div className="mt-2">
-                  <InlineNotification
-                    type="info"
-                    desc={t('insufficient-sol')}
+                <div className="flex items-center justify-between rounded-md border border-th-bkg-3 px-3 py-2">
+                  <div>
+                    <p className="text-th-fgd-2">{t('enable-notifications')}</p>
+                    <p className="text-xs">{t('asked-sign-transaction')}</p>
+                  </div>
+                  <Switch
+                    className="text-th-fgd-3"
+                    checked={signToNotifications}
+                    onChange={(checked) => setSignToNotifications(checked)}
                   />
+                </div>
+                <div className="space-y-3">
                   <div className="mt-10">
                     <Button
                       className="mb-6 flex items-center justify-center"
