+++ public/locales/ru/common.json
@@ -19,6 +19,7 @@
   "all": "All",
   "amount": "Amount",
   "amount-owed": "Amount Owed",
+  "asked-sign-transaction": "You'll be asked to sign a transaction",
   "asset-liability-weight": "Asset/Liability Weights",
   "asset-liability-weight-desc": "Asset weight applies a haircut to the value of the collateral in your account health calculation. The lower the asset weight, the less the asset counts towards collateral. Liability weight does the opposite (adds to the value of the liability in your health calculation).",
   "asset-weight": "Asset Weight",
@@ -75,6 +76,7 @@
   "edit": "Edit",
   "edit-account": "Edit Account Name",
   "edit-profile-image": "Edit Profile Image",
+  "enable-notifications": "Enable Notifications",
   "error-borrow-exceeds-limit": "Maximum borrow for the current period is {{remaining}}. New period starts {{resetTime}}",
   "error-token-positions-full": "No token positions available in your account.",
   "explorer": "Explorer",
@@ -194,6 +196,7 @@
   "list-market-token": "List Market/Token",
   "vote": "Vote",
   "yes": "Yes",
-  "you": "You"
+  "you": "You",
+  "using-ledger": "Using Ledger"
  }
