+++ public/locales/zh_tw/common.json
@@ -19,6 +19,7 @@
   "all": "全部",
   "amount": "數量",
   "amount-owed": "欠款",
+  "asked-sign-transaction": "You'll be asked to sign a transaction",
   "asset-liability-weight": "資產/債務權重",
   "asset-liability-weight-desc": "資產權重在賬戶健康計算中對質押品價值進行扣減。資產權重越低，資產對質押品的影響越小。債務權重恰恰相反（在健康計算中增加債務價值）。",
   "asset-weight": "資產權重",
@@ -75,6 +76,7 @@
   "edit": "編輯",
   "edit-account": "編輯帳戶標籤",
   "edit-profile-image": "切換頭像",
+  "enable-notifications": "Enable Notifications",
   "error-borrow-exceeds-limit": "Maximum borrow for the current period is {{remaining}}. New period starts {{resetTime}}",
   "error-token-positions-full": "No token positions available in your account.",
   "explorer": "瀏覽器",
@@ -193,5 +195,6 @@
   "list-market-token": "列出市場／幣種",
   "vote": "投票",
   "yes": "是",
-  "you": "你"
+  "you": "你",
+  "using-ledger": "Using Ledger"
 }
